12. EntityGraph
Using EntityGraph to select relations dynamically with AttributeNodes.

With this functionality we could use an entity with LAZY Relations and fetch then dynamically using AttributeNodes of EntityGraph.

We used 2 types of graphs:

fetchgraph: will select only the attributes specified ignoring the others EAGERS on entity, even explicit.

loadgraph: will select the attributes specified and the others EAGERS defined on entity.